You don't have to keep track of the loop counter yourself. Let python do it for you. It's good at it.
I'm using xrange here because in python 2.* range creates a list of all
the numbers it will go through and then iterates over it, so if you enter
a trillion, it will use up all the memory on your computer storing the list,
and your computer will die. I don't know why. This seems stupid to me.

xrange just creates an iterator (basically, it counts to the number you
give it,) so it might take a long time, but it wil actually work.

In python 3.* this is fixed.